chore(test): move mock file to bootstrap file.
